pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["NICOLAS HAVARD <nicolas.havard.contractor@ardian.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
weasyprint = "^63.1"
pypdf = "^5.1.0"
pdfrw = "^0.4"
pyhanko = {extras = ["opentype"], version = "^0.25.3"}
jinja2 = "^3.1.5"
shapely = "^2.0.6"
pyproj = "^3.7.0"
requests = "^2.32.3"
django = "^5.1.7"
djangorestframework = "^3.15.2"
django-environ = "^0.12.0"
djangorestframework-gis = "^1.1"
owslib = "^0.33.0"
fiona = "^1.10.1"
pandas = "^2.2.3"
odfpy = "^1.4.1"
django-cors-headers = "^4.7.0"
pymupdf = "^1.25.5"
python-slugify = "^8.0.4"
gunicorn = "^23.0.0"
factory-boy = "^3.3.3"
whitenoise = "^6.9.0"
django-ratelimit = "^4.1.0"
django-redis = "^5.4.0"
djangorestframework-simplejwt = "^5.5.0"
google-auth = "^2.40.1"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.1"
num2words = "^0.5.14"
pillow = "^11.3.0"
drf-spectacular = "^0.28.0"
django-simple-history = "^3.10.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.11.1"
faker = "^37.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88  # Comme Black
target-version = "py311"  # Change selon ta version Python

[tool.ruff.lint]
select = ["E", "F", "W"]  # Active seulement les règles PEP8, Flake8, et warnings

[tool.ruff.format]
quote-style = "double"  # Convertit les guillemets simples en doubles
indent-style = "space"  # Utilise des espaces au lieu de tabs
skip-magic-trailing-comma = false  # Laisse la virgule finale
